Refactor engine.
Split engine out into packages
Split room definition out into packages
Added ItemDescriptionHandlers
Made parsing logic common parent base of roomcommands/itemcommands
Added Exits
Rebuilt WebSocket registration to use programmatic endpoint creation
instead of annotatated. 
Removed state concept from Exits.
