+++ concierge-app/src/main/java/net/wasdev/gameon/concierge/PlacementStrategy.java
@@ -1,8 +1,5 @@
 package net.wasdev.gameon.concierge;
 
-import java.util.List;
-import java.util.UUID;
-
 import net.wasdev.gameon.room.common.Room;
 
 public interface PlacementStrategy {
