+++ room-app/src/main/java/net/wasdev/gameon/room/engine/DataProvider.java
@@ -1,0 +1,7 @@
+package net.wasdev.gameon.room.engine;
+
+import java.util.Collection;
+
+public interface DataProvider {
+	Collection<Room> getRooms();
+}
